[MOD] Pickup when function from support collection to whenable trait
