+++ src/Illuminate/Support/Traits/Whenable.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Illuminate\Support\Traits;
+
+trait Whenable
+{
+    /**
+     * Apply the callback if the value is truthy.
+     *
+     * @param  bool  $value
+     * @param  callable  $callback
+     * @param  callable  $default
+     * @return mixed
+     */
+    public function when($value, callable $callback, callable $default = null)
+    {
+        if ($value) {
+            return $callback($this);
+        } elseif ($default) {
+            return $default($this);
+        }
+
+        return $this;
+    }
+}
